Link insertion with no selected text
TinyMCE requires text to be selected in order to insert links.
Added checks for if there is a text selection, and insert HTML content if there is not.
Link contents consists of the target name, or the target url if not populated

(cherry picked from commit e7f8e692c2397f268e864524b892b65acff3d29a)