Apply pickDate-config setting to hide calendar in datepicker (#11632)
* Apply pickDate-config setting to hide calendar in datepicker 

A time-only picker is only possible if the format is configured like "hh:mm:ss". By using the pickDate config it is possible to determine whether the calendar should be shown or not, instead of depending on a certain time format.

* Changed timeformat to accept both hh:mm:ss and hh:mm

Edited the regex to math times with and without seconds. The calendar should not be shown anymore if the timeformat is (H)H:mm:ss or (H)H:mm.

* Removed unused config property pickDate

config.pickDate is never used and can be removed.

* remove pickDate config from datepickers config