move code-block to backoffice folder
Tests are failing when autoimporting an index file. As it turned out, the 'utils' folder is not the right place to put shared elements. Instead we are creating a 'shared' folder in the backoffice context and autoloading the component globally for the backoffice.

Consider moving the code-block element to the UI library later.