#14325 - Remove wildcard-background from RTE (#14439)
* Update rte-content.less

#14325 - Remove wildcard-background from RTE

* Update rte-content.less

I did an update. 🙂

The transparent background is there because the tinymce theme Oxide adds a background-color to all selected elements.
So the umbraco added transparent background makes white text not be visible.

https://github.com/tinymce/oxide/blob/012346905578b0d7e59f7000d447f56243243b4a/src/less/theme/content/selection/selection.less#L50C2-L50C2


So I put back the transparent background-color and added a @blueExtraDark as text-color

I have updated the issue with more details.