Media Picker - upload button disabled bug #13383
Issue is around how the folder path is "got" on the initialisation - if it doesn't come from localstorage (which is wiped on logout- hence how I could recreate). The function gotoFolder(folder) has entityResource.getAncestors that builds the folder path if it's not set (which it isn't on first init) but this isn't on a promise or awaited.

Modified the code that follows so it can only run once this returns by moving it to a separate function performGotoFolder(folder). It then has the path and on first open the editor can add images.

I've tested the following scenarios and it seems to work (without breaking everything).

Media picker with start node set.
Media picker without start node
User with limited media folders (e.g. can't see the start node).
Seems to have resolved the issue.