15209 media picker shows incorrect date (#15414)
* Trim update date to exclude Z character.

When picking media, the items from an SQL database contain an update
date that is formated in UTC-0 time by way of a Z character at the end
of the string. Removing this string will result in the update date
always displaying local time as expected.

Issue: 15209

* Use dateHelper instead of slicing Z from date string.

For most front end date formatting, dateHelper is used to get the local
date. The dateHelper is now passed in to the media picker controller and
used to format the date.