Unable to correct an invalid alias on a property
When adding a new property with a short name such as A, an invalid alias will be created.

If you then correct just the alias you won't be able to save your new property.  I've added a Playwright test which demos this.

You can get around this by changing the name and the alias but it isn't immediately obvious, as the error on the name isn't flagged.