Corrected casing of `Umbraco.MultipleTextString`
Whilst testing the property-editors, I'd noticed that the
Multiple Textstring editor wasn't loading correctly, this is because
the schema alias on the server is `Umbraco.MultipleTextstring`
(lowercase 's') and on the client it is `Umbraco.MultipleTextString`
(uppercase 'S'). I've corrected this to align with the server.

I assume that the intent was to lowercase the 'S' on the server,
but this didn't happen before the beta release.