fix: Tinymce allows potentially unsafe embeds
This sets `convert_unsafe_embeds` to true to act as a workaround for CVE-2024-29881 seeing that we cannot upgrade to tinymce 7 at the moment. The only difference from TinyMCE 6 to 7 in this regard is that this option is set to true and is the official workaround.

Fixes https://github.com/umbraco/Umbraco.CMS.Backoffice/security/dependabot/44