fix: prevent copying of the node_modules directory
As noted in #16659, everything from `App_Plugins` is copied into the web
output. This is unneeded and in the case of the `node_modules` directory
unwanted behavior, these files are sources and not output that Umbracco
needs to serve packages/extensions.

The old behavior has led to many of the "path too long" errors we've
been seeing while building extensions.