Updated log viewer date range selector
Added theme property to the log viewer date range selector component. This allows for dynamic theme changes in the application. Also, added a new import for UMB_THEME_CONTEXT and used it within the constructor to observe any changes in the theme context. Additionally, updated CSS styles to invert color of calendar picker indicator when dark theme is active.