fix: pins the UmbracoExtension `@hey-api/*` versions to that of the Backoffice client (#20016)
fix: pins the @hey-api/* versions to that of the Backoffice client

This is a quick fix to handle the NPM error that is currently there because the Backoffice NPM client has moved on to another version. There will be a more comprehensive fix for 16.3, however this PR aims to make the 16.2 UmbracoExtension usable without running custom commands.