Change ExamineIndexRebuilder to public access modifier
Without this being public it is not possible to override anything or resolve this internally without some ugly reflection being done at the composition layer. ExamineX has had to copy this internally to override some methods since index rebuilding should not be done for certain server roles. Similarly, for persisted indexes in managed search services like ExamineX with Azure or Elasticsearch, rebuilding on cold boot should not be done so this behavior needs to be overridden.