Fixes the warning CS8524 and tidying up the SetStatusRedirectUrlManagementController
I have added throwing an error in the switch statement if neither of the know enums is submitted to the API, in theory this should never happen.

The other option would be to simply return false, but I think throwing the exception is better as it will alert whoever is calling the API that their call is invalid.