Fixing the build warnings in the Umbraco.Tests.Common Project
The main warning was about the file name not matching the class, this was because an interface was being defined first within the class file.

This should either be moved into its own file, or to the bottom of the file to fix this warning. Rather than creating a new file, I've moved the interface to the bottom, but if you'd prefer a new file, just let me know :)

Also removed the TODO in the project file and the WarningsNotAsErrors as they have all been resolved.